+++ lib/select_chips_input.dart
@@ -1,3 +1,3 @@
-library select_chips_input.dart;
+library select_chips_input;
 
 export 'src/select_chips_input.dart';